+++ R/Subset.R
@@ -1,15 +1,15 @@
-#' @title Subsets a data frame and drops the unused levels.
+#' @name Subset
+#' 
+#' @title Subsets/filters a data frame and drops the unused levels.
 #'
-#' @description Subsets a data frame and drops the unused levels.
+#' @description Subsets/filters a data frame and drops the unused levels.
 #'
-#' @details This function is used only for data frames.  It is equivalent to the combined usage of \code{\link{subset}} and \code{\link{drop.levels}}.  Use \code{\link{subset}} for all other structures.
-#'
-#' @note Newbie students using R expect that when a factor variable is subsetted with \code{subset} that any original levels that are no longer used after the subsetting will be ignored.  This, however, is not the case and often results in tables with empty cells and figures with empty bars.  One remedy is to use \code{drop.levels} immediately following the \code{subset} call.  This generally becomes a repetitive sequence for most newbie students; thus, \code{Subset} incorporates these two functions into one function.
-#'
-#' @note This is the code from \code{\link{subset}} with a catch for non-data.frames and a specific call to \code{drop.levels} just before the data.frame is returned.  I also added an argument to allow resetting the row names.  Otherwise, there is really no new code here.
-#'
-#' @aliases Subset Subset.data.frame
-#'
+#' @details Newbie students using R expect that when a factor variable is subsetted with \code{\link{subset}} or filtered with \code{\link[dplyr]{filter}} that any original levels that are no longer used after the subsetting or flitering will be ignored.  This, however, is not the case and often results in tables with empty cells and figures with empty bars.  One remedy is to use \code{\link[gdata]{drop.levels}} from \pkg{gdata} immediately following the \code{\link{subset}} or \code{\link[dplyr]{filter}} call.  This generally becomes a repetitive sequence for most newbie students; thus, \code{Subset} and \code{filterD} incorporate these two functions into one function.
+#' 
+#' \code{Subset} is a wrapper to \code{\link{subset}} with a catch for non-data.frames and a specific call to \code{\link[gdata]{drop.levels}} just before the data.frame is returned.  I also added an argument to allow resetting the row names.  \code{filterD} is a wrapper for \code{\link[dplyr]{filter}} from \pkg{dplyr} followed by \code{\link[gdata]{drop.levels}} just before the data.frame is returned.  Otherwise, there is no new code here.
+#' 
+#' These functions are used only for data frames.
+#' 
 #' @param x A data frame.
 #' @param subset A logical expression that indicates elements or rows to keep: missing values are taken as false.
 #' @param select An expression, that indicates columns to select from a data frame.
@@ -20,8 +20,10 @@
 #' @return A data frame with the subsetted rows and selected variables.
 #'
 #' @author Derek H. Ogle, \email{dogle@@northland.edu}
+#' 
+#' @section IFAR Chapter: \href{https://fishr.wordpress.com/books/ifar/}{2-Basic Data Manipulations}.
 #'
-#' @seealso See \code{subset}, \code{\link[dplyr]{filter}} from \pkg{dplyr}, and \code{\link{filterD}} for similar functionality.  See \code{\link[gdata]{drop.levels}} in \pkg{gdata} for related functionality.
+#' @seealso See \code{subset} and \code{\link[dplyr]{filter}} from \pkg{dplyr} for similar functionality.  See \code{\link[gdata]{drop.levels}} in \pkg{gdata} and \code{\link{droplevels}} for related functionality.
 #'
 #' @keywords misc
 #'
@@ -30,23 +32,24 @@
 #' levels(iris$Species)
 #' iris.set1 <- subset(iris,Species=="setosa" | Species=="versicolor")
 #' levels(iris.set1$Species)
-#' table(iris.set1$Species)
+#' xtabs(~Species,data=iris)
 #'
 #' ## A simpler fix using Subset
 #' iris.set2 <- Subset(iris,Species=="setosa" | Species=="versicolor")
 #' levels(iris.set2$Species)
-#' table(iris.set2$Species)
+#' xtabs(~Species,data=iris.set2)
+#' 
+#' ## A simpler fix using filterD
+#' iris.set3 <- filterD(iris,Species=="setosa" | Species=="versicolor")
+#' levels(iris.set3$Species)
+#' xtabs(~Species,data=iris.set3)
 #'
-#' @rdname Subset
-#' @export
-Subset <- function (x,...) {
-  if (!is.data.frame(x)) stop("Subset should only be used with data frames.  See ?subset if using a different structure.",call.=FALSE)
-  UseMethod("Subset") 
-}
+NULL
 
 #' @rdname Subset
 #' @export
-Subset.data.frame <- function (x,subset,select,drop=FALSE,resetRownames=TRUE,...) {
+Subset <- function(x,subset,select,drop=FALSE,resetRownames=TRUE,...) {
+  if (!is.data.frame(x)) stop("Subset should only be used with data frames.  See ?subset for other structures.",call.=FALSE)
   if (missing(subset)) r <- TRUE
   else {
     e <- substitute(subset)
@@ -60,8 +63,17 @@ Subset.data.frame <- function (x,subset,select,drop=FALSE,resetRownames=TRUE,...
     names(nl) <- names(x)
     vars <- eval(substitute(select),nl,parent.frame())
   }
-  res <- drop.levels(x[r,vars,drop = drop],reorder=FALSE)
+  res <- gdata::drop.levels(x[r,vars,drop = drop],reorder=FALSE)
   if (resetRownames) rownames(res) <- NULL
+  if (nrow(res)==0) warning("The resultant data.frame has 0 rows.  Try str() on the result.\n",call.=FALSE)
+  res
+}
+
+#' @rdname Subset
+#' @export
+filterD <- function(x,...) {
+  res <- dplyr::filter(x,...)
+  res <- gdata::drop.levels(res)
   if (nrow(res)==0) warning("The resultant data.frame has 0 rows.  Try str() on the result.\n",call.=FALSE)
   res
 }